Update README.md
# AndroidGameDatabase
I want to build a database that has as much information about each game as possible. 
Although many apps and sites do this to some extent, there always seems to be some important piece of info 
that has to be obtained elsewhere. I also wanted to have a more specific organizational option,
 such as graphics type, gameplay type or similar games, year and platform releases etc. With the ultimate
Goal of being able to have only the games your interested in or looking for after your search specification 
The are too many games in the playstore that I don't want to see show up in searches. I also wanted to add a few 
special search options that can show only your favorite types of games as well as a favorites list with games 
you want from any location. 
My bullet point list of proposed features would be:
1 can choose your oun and add, any search engine or apps and tools to search with
2 can add any links to each game profile to DL the game, gameplay pics, plugins/mods etc
3 each use can build a smaller more specific database for themselves for future use